implement mixed History
